move header tabs to component; remove contentclass from App component
