+++ public/styles.css
@@ -94,6 +94,10 @@ h4 {
   color: var(--color-text-light-grey);
 }
 
+.align-right {
+  margin-left: auto;
+}
+
 @media (max-width: 580px) {
   table, thead, tbody, th, td, tr {
     display: block;
