+++ src/App.svelte
@@ -3,6 +3,7 @@
   import { Student, ProcessToken, NotFound, Login } from "./routes";
   import { Badges, Issuers } from "./routes/teachers";
   import { Header, Footer } from "./components";
+  import { Header as TeacherHeader } from "./components/teachers";
   import { userRole, userLoggedIn } from "./stores/user";
   import { role } from "./util/role";
 </script>
@@ -17,6 +18,8 @@
     flex-direction: column;
 
     box-shadow: 0 0 8px rgba(0, 0, 0, 0.4);
+
+    background-color: white;
   }
 
   @media (min-width: 600px) {
@@ -24,58 +27,44 @@
       margin: 0 40px;
     }
   }
-
-  .content {
-    flex: 1;
-    background-color: white;
-
-    display: flex;
-  }
 </style>
 
 <div class="app">
   {#if $userLoggedIn && $userRole === role.STUDENT}
     <Router>
-      <Header />
-
-      <div class="content">
-        <Route path="/" component={Student} />
-        <Route path="/backpack">
-          <Student bookmark="backpack" />
-        </Route>
-        <Route path="/badge-requests">
-          <Student bookmark="badge-requests" />
-        </Route>
-        <Route path="/collections">
-          <Student bookmark="collections" />
-        </Route>
-        <Route path="/profile">
-          <Student bookmark="profile" />
-        </Route>
-        <Route path="/auth/login/*" component={ProcessToken} />
-        <Route component={NotFound} />
-      </div>
+      <Header logout />
+      <Route path="/" component={Student} />
+      <Route path="/backpack">
+        <Student bookmark="backpack" />
+      </Route>
+      <Route path="/badge-requests">
+        <Student bookmark="badge-requests" />
+      </Route>
+      <Route path="/collections">
+        <Student bookmark="collections" />
+      </Route>
+      <Route path="/profile">
+        <Student bookmark="profile" />
+      </Route>
+      <Route path="/auth/login/*" component={ProcessToken} />
+      <Route component={NotFound} />
     </Router>
   {:else if $userLoggedIn && $userRole === role.TEACHER}
     <Router>
-      <Header tabs={['/', '/issuers']} />
+      <TeacherHeader />
 
-      <div class="content">
-        <Route path="/" component={Badges} />
-        <Route path="/issuers" component={Issuers} />
-        <Route path="/auth/login/*" component={ProcessToken} />
-        <Route component={NotFound} />
-      </div>
+      <Route path="/" component={Badges} />
+      <Route path="/issuers" component={Issuers} />
+      <Route path="/auth/login/*" component={ProcessToken} />
+      <Route component={NotFound} />
     </Router>
   {:else}
     <Router>
       <Header />
 
-      <div class="content">
-        <Route path="/" component={Login} />
-        <Route path="/auth/login/*" component={ProcessToken} />
-        <Route component={NotFound} />
-      </div>
+      <Route path="/" component={Login} />
+      <Route path="/auth/login/*" component={ProcessToken} />
+      <Route component={NotFound} />
     </Router>
   {/if}
