+++ src/components/index.js
@@ -1,10 +1,10 @@
-export { default as Badge } from "./Badge.svelte"
-export { default as Button } from "./Button.svelte"
-export { default as CheckBox } from "./CheckBox.svelte"
-export { default as FilterBlock } from "./FilterBlock.svelte"
-export { default as Footer } from "./Footer.svelte"
-export { default as Header } from "./Header.svelte"
-export { default as Layout } from "./Layout.svelte"
-export { default as Search } from "./Search.svelte"
-export { default as SideBar } from "./SideBar.svelte"
-export { default as SideMenu } from "./SideMenu.svelte"
+export { default as Badge } from "./Badge.svelte";
+export { default as Button } from "./Button.svelte";
+export { default as CheckBox } from "./CheckBox.svelte";
+export { default as FilterBlock } from "./FilterBlock.svelte";
+export { default as Footer } from "./Footer.svelte";
+export { default as Header } from "./Header.svelte";
+export { default as Layout } from "./Layout.svelte";
+export { default as Search } from "./Search.svelte";
+export { default as SideMenu } from "./SideMenu.svelte";
+export { default as LogoutButton } from "./LogoutButton.svelte";
