+++ src/components/teachers/SideBar.svelte
@@ -1,6 +1,6 @@
 <script>
-  import { FilterBlock, Search } from "../components";
-  import { tree, facultyIds, issuerIds } from "../stores/filter";
+  import { FilterBlock, Search } from "../../components";
+  import { tree, facultyIds, issuerIds } from "../../stores/filter";
 </script>
 
 <style>
