+++ src/components/teachers/index.js
@@ -1,0 +1,2 @@
+export { default as Header } from "./Header.svelte";
+export { default as SideBar } from "./SideBar.svelte";
